pyproject.toml
[project]
name = 'MQL-Back'
version = '1.0.0'
description = 'Backend for the MQL Trading Bot'
requires-python = '>=3.11'
dependencies = [
    'alembic',
    'psycopg2-binary',
    'Flask',
    'flask-blueprint',
    'Flask-Cors',
    'Flask-Migrate',
    'marshmallow',
    'marshmallow-sqlalchemy',
    'requests',
    'SQLAlchemy',
    'uWSGI',
    'load_dotenv',
    'pytest',
    'Flask-JWT-Extended',
    'selenium',
    'bs4',
    'webdriver-manager',
    'apscheduler',
]

[pyproject.optional-dependencies]
dev = [
    'debugpy'
]

[tool.setuptools]
py-modules = []

[tool.black]
line-length = 120

[tool.isort]
profile = 'black'